test(http): keep rejection fixtures typed and scoped
Prove queued responses exist before use and move the abort-activity test into its own file with shared signing/account fixtures. Preserve all wire assertions and clarify the Node HEAD callback contract.
